fix: separate resname tokenizing from class composer
